+++ hosting/src/entities/record.ts
@@ -1,26 +1,19 @@
 import { calculateVDOT } from '../utils/vdot'
+
 export interface Record {
   key: string
   seconds: number
   meters: number
   date: firebase.firestore.Timestamp
-  vdot(): number
 }
 
-export class Record implements Record {
-  constructor(data: {
-    key: string
-    seconds: number
-    meters: number
-    date: firebase.firestore.Timestamp
-  }) {
-    this.key = data.key
-    this.seconds = data.seconds
-    this.meters = data.meters
-    this.date = data.date
-
-    console.log(this.seconds)
+export class Record {
+  constructor(r: Record) {
+    this.key = r.key
+    this.seconds = r.seconds
+    this.meters = r.meters
+    this.date = r.date
   }
 
-  public vdot = () => calculateVDOT(this.seconds, this.meters)
+  public vdot = (): number => calculateVDOT(this.seconds, this.meters)
 }
